Git: Improve the default for 'git.main_branch'
If the option is unset, the main branch is set from the HEAD of the main
remote, during the initial checkout. The main branch is not set at all
if the main remote is not correctly configured.

Instead of defaulting to "master", which might lead to failure during
the initial checkout of repositories that use an other default branch.